rename state files
