+++ src/app/state/configAtom.ts
