+++ src/app/state/index.ts
@@ -1,6 +1,6 @@
-export { default as credentialAtom } from "./credenial";
-export { default as configAtom } from "./config";
-export { default as signedInSelector } from "./signedIn";
-export { default as currentUserSelector } from "./currentUser";
-export { default as mockedSelector } from "./mocked";
-export { default as loginDisabledSelector } from "./loginDisabled";
+export { default as credentialAtom } from "./credenialAtom";
+export { default as configAtom } from "./configAtom";
+export { default as signedInSelector } from "./signedInSelector";
+export { default as currentUserSelector } from "./currentUserSelector";
+export { default as mockedSelector } from "./mockedSelector";
+export { default as loginDisabledSelector } from "./loginDisabledSelector";
